+++ core/src/sn/TxtStage.ts
@@ -5,8 +5,6 @@
 	http://opensource.org/licenses/mit-license.php
 ** ***** END LICENSE BLOCK ***** */
 
-import {Container, Texture, Sprite, Graphics, Rectangle, Renderer} from 'pixi.js';
-
 import {CmnLib, IEvtMng, argChk_Boolean, argChk_Num} from './CmnLib';
 import {HArg} from './CmnInterface';
 import {Config} from './Config';
@@ -14,7 +12,8 @@ import {CmnTween} from './CmnTween';
 import {GrpLayer} from './GrpLayer';
 import {DebugMng} from './DebugMng';
 
-const Tween = require('@tweenjs/tween.js').default;
+import {Container, Texture, Sprite, Graphics, Rectangle, Renderer} from 'pixi.js';
+import {Tween} from '@tweenjs/tween.js'
 
 export interface IInfTxLay {
 	fontsize	: number;
@@ -37,7 +36,7 @@ interface IChRect {
 }
 interface ISpTw {
 	sp	: Container;
-	tw	: TWEEN.Tween | null;
+	tw	: Tween<Container> | null;
 };
 
 export class TxtStage extends Container {
@@ -770,7 +769,7 @@ export class TxtStage extends Container {
 		);
 		const st: ISpTw = {
 			sp: sp,
-			tw: new Tween.Tween(sp)
+			tw: new Tween(sp)
 				.to({ alpha: 1, x: rct.x, y: rct.y, width: rct.width, height: rct.height, angle: 0 }, cis.wait ?? 0)
 				.easing(ease)
 				.delay((add.wait ?? 0) +(arg.delay ?? 0))